refactor(tools): deduplicate fetch_issues; pin workflow action SHAs
Move fetch_issues() to sonarcloud_constants so sync and triage
share one implementation. Pin actions/checkout and astral-sh/setup-uv
to commit SHAs per supply-chain security guidance.
